// src/app/books/books.types.ts
export type Book = {
    _id: string
    name: string
    price: string
    description: string
    category: string
    createdAt: string
    image: string
}

export type Cart = {
    _id: string
    name: string
    price: string
    description: string
    category: string
    createdAt: string
    quantity: number
    image: string
}